Move TaskUtil to util.conversion package
