+++ app/src/androidTest/java/com/groupseven/musicmap/util/conversion/TaskUtilTest.java
@@ -1,14 +1,17 @@
-package com.groupseven.musicmap.util;
+package com.groupseven.musicmap.util.conversion;
 
-import com.google.android.gms.tasks.Task;
-import com.google.android.gms.tasks.Tasks;
-import org.junit.Test;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
+
+import com.google.android.gms.tasks.Task;
+import com.google.android.gms.tasks.Tasks;
+
+import org.junit.Test;
+
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
 
 public class TaskUtilTest {
