+++ app/src/main/java/com/groupseven/musicmap/util/firebase/Actions.java
@@ -9,7 +9,7 @@ import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageException;
 import com.google.firebase.storage.StorageReference;
 import com.groupseven.musicmap.models.MusicMemory;
-import com.groupseven.musicmap.util.TaskUtil;
+import com.groupseven.musicmap.util.conversion.TaskUtil;
 
 import java.io.ByteArrayOutputStream;
 import java.util.UUID;
