+++ app/src/main/java/com/groupseven/musicmap/util/firebase/AuthSystem.java
@@ -24,7 +24,7 @@ import com.google.firebase.firestore.WriteBatch;
 import com.google.firebase.internal.api.FirebaseNoSignedInUserException;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
-import com.groupseven.musicmap.util.TaskUtil;
+import com.groupseven.musicmap.util.conversion.TaskUtil;
 
 import java.util.HashMap;
 import java.util.Map;
